Breaks purgecss into modules
